Simplified plotting. Use pathlib. Add regression test for convection-coupled binary alloy solidification. Benchmark regression test now fails
